fix output directory bug to go to version 1.01

+++ src/main/java/org/eumetsat/eoportal/dcpc/commons/FileSystem.java
@@ -62,17 +62,20 @@ public class FileSystem
 
             File[] files = path.listFiles();
 
-            for (int i = 0; i < files.length; ++i)
+            if (files != null)
             {
-                if (files[i].isDirectory())
+                for (int i = 0; i < files.length; ++i)
                 {
-                    deleteDirs(files[i]);
+                    if (files[i].isDirectory())
+                    {
+                        deleteDirs(files[i]);
+                    }
+                    files[i].delete();
                 }
-                files[i].delete();
             }
             path.delete();
         }
-        catch (Exception ignored)
+        catch (Throwable ignored)
         {
             logger.error("Cannot deleteDirs ", ignored);
         }
